Plane: convert distance params from AP_Int16 to AP_Float for GCS unit conversion
Fixes #32606

RTL_RADIUS, WP_RADIUS, WP_MAX_RADIUS and WP_LOITER_RAD were declared
as AP_Int16 which prevents GCS applications from applying unit
conversion (e.g. meters to feet). The MAVLink protocol already sends
all parameters as floats, so this change has no wire format impact.

Also removed an explicit uint16_t cast in commands_logic.cpp that
became unnecessary after the type change.

+++ ArduPlane/commands_logic.cpp
@@ -660,7 +660,7 @@ bool Plane::verify_nav_wp(const AP_Mission::Mission_Command& cmd)
     // see if the user has specified a maximum distance to waypoint
     // If override with p3 - then this is not used as it will overfly badly
     if (g.waypoint_max_radius > 0 &&
-        auto_state.wp_distance > (uint16_t)g.waypoint_max_radius) {
+        auto_state.wp_distance > g.waypoint_max_radius) {
         if (current_loc.past_interval_finish_line(prev_WP_loc, flex_next_WP_loc)) {
             // this is needed to ensure completion of the waypoint
             if (cmd_passby == 0) {
